src/main/webapp/app/entities/dispositif-signature/delete/dispositif-signature-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DispositifSignatureService } from '../service/dispositif-signature.service';

import { DispositifSignatureDeleteDialogComponent } from './dispositif-signature-delete-dialog.component';

describe('DispositifSignature Management Delete Component', () => {
  let comp: DispositifSignatureDeleteDialogComponent;
  let fixture: ComponentFixture<DispositifSignatureDeleteDialogComponent>;
  let service: DispositifSignatureService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DispositifSignatureDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DispositifSignatureDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DispositifSignatureDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DispositifSignatureService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
